pyproject.toml
[tool.poetry]
name = "musicbot"
version = "0.1.0"
description = ""
authors = ["Jorge <jsjfuentesj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.3"
discord-py = {extras = ["voice"], version = "^2.1.0"}
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
